Initialize a targets project

Initialize a targets project

tar_init(
  home = ".",
  phase_names = c("fetch", "process", "summarize"),
  phase_nums = seq_along(phase_names),
  separate_phase_scripts = TRUE,
  phase_subdirs = c("src", "out"),
  overwrite = FALSE
)

Arguments

home

char, root directory of targets project. Defaults to current working directory

phase_names

char vector, names of target phases

phase_nums

int vector, numbers to prepend to phase_names like "1_fetch", "2_process", etc. Defaults to 1:length(phase_names)

separate_phase_scripts

lgl, should a different R script be created for each phase like "1_fetch.R", "2_process.R", etc? If FALSE, then targets lists will be initialized in _targets.R file

phase_subdirs

char vector, subdirectories within each phase like "src", "out"

overwrite

lgl, should the initialization overwrite files and folders that already exist?

Value

NULL invisibly