Add individual USGS project files to a directory

Creates common project files like .gitignore, README.md. LICENSE.md, etc. based on the templates here: https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/.

use_gitignore_usgs(
  home = ".",
  additions = NULL,
  open = rlang::is_interactive()
)

use_readme_usgs(home = ".", open = rlang::is_interactive())

use_readme_rmd_usgs(home = ".", open = rlang::is_interactive())

use_license_usgs(home = ".", open = rlang::is_interactive())

use_disclaimer_provisional_usgs(home = ".", open = rlang::is_interactive())

use_disclaimer_approved_usgs(home = ".", open = rlang::is_interactive())

use_code_of_conduct_usgs(home = ".", open = rlang::is_interactive())

use_contributing_usgs(home = ".", open = rlang::is_interactive())

use_changelog_usgs(home = ".", open = rlang::is_interactive())

Arguments

home

chr, root directory of project. Defaults to current working directory

additions

chr vector, other files/directories to be added to .gitignore

open

lgl; whether to open the file for interactive editing

Examples

tmp <- tempdir()

use_gitignore_usgs(home = tmp,
                   additions = c("excluded_file.R",
                                 "excluded_dir",
                                 "*excluded_pattern*"))
#> • Edit 'REDACTED FILE PATH'
#> ℹ Using .gitignore from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>

# here are the contents of the .gitignore:
cat(readLines(file.path(tmp,".gitignore")), sep = "\n")
#> # **NOTE: this is a template created using the `mortar` package. The contents below are for illustrative purposes and should be edited to reflect your use case. This message can be deleted in acknowledgement that you've added this file to your project.**
#> 
#> # Copied from https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/-/blob/master/.gitignore?ref_type=heads
#> 
#> ### R ###
#> 
#> # R Environment Variables
#> .Renviron
#> 
#> ### targets ###
#> _targets
#> */out/*
#> */in/*
#> !*.empty
#> 
#> # History files
#> .Rhistory
#> .Rapp.history
#> .RData # Session Data files
#> .Ruserdata # User-specific files
#> .Rproj.user/ # RStudio files
#> 
#> # OAuth2 token, see https://github.com/hadley/httr/releases/tag/v0.3
#> .httr-oauth
#> 
#> # knitr and R markdown default cache directories
#> *_cache/
#> /cache/
#> 
#> # Temporary files created by R markdown
#> *.utf8.md
#> *.knit.md
#> 
#> ### vim ###
#> .*.swp
#> .swp
#> excluded_file.R
#> excluded_dir
#> *excluded_pattern*