Create a .gitignore file with an opinionated list of ignores

Uses .gitignore file from https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/. Specify project-specific exclusions with the additions argument, which are appended to the end of the .gitignore.

use_gitignore_usgs(home = ".", additions = NULL)

Arguments

home

chr, root directory of project. Defaults to current working directory

additions

chr vector, other files/directories to be added to .gitignore

Examples

tmp <- tempdir()
use_gitignore_usgs(home = tmp,
                   additions = c("excluded_file.R",
                                 "excluded_dir",
                                 "*excluded_pattern*"))
#> ℹ Using .gitignore from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>

# here are the contents of the .gitignore:
cat(readLines(file.path(tmp,".gitignore")), sep = "\n")
#> # Copied from https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/-/blob/master/.gitignore?ref_type=heads
#> # Created by https://www.toptal.com/developers/gitignore/api/r,python,jupyternotebooks,vscode
#> # Edit at https://www.toptal.com/developers/gitignore?templates=r,python,jupyternotebooks,vscode
#> 
#> ### JupyterNotebooks ###
#> # gitignore template for Jupyter Notebooks
#> # website: http://jupyter.org/
#> 
#> .ipynb_checkpoints
#> */.ipynb_checkpoints/*
#> 
#> # IPython
#> profile_default/
#> ipython_config.py
#> 
#> # Remove previous ipynb_checkpoints
#> #   git rm -r .ipynb_checkpoints/
#> 
#> ### Python ###
#> # Byte-compiled / optimized / DLL files
#> __pycache__/
#> *.py[cod]
#> *$py.class
#> 
#> # C extensions
#> *.so
#> 
#> # Distribution / packaging
#> .Python
#> build/
#> develop-eggs/
#> dist/
#> downloads/
#> eggs/
#> .eggs/
#> parts/
#> sdist/
#> var/
#> wheels/
#> pip-wheel-metadata/
#> share/python-wheels/
#> *.egg-info/
#> .installed.cfg
#> *.egg
#> MANIFEST
#> 
#> # PyInstaller
#> #  Usually these files are written by a python script from a template
#> #  before PyInstaller builds the exe, so as to inject date/other infos into it.
#> *.manifest
#> *.spec
#> 
#> # Installer logs
#> pip-log.txt
#> pip-delete-this-directory.txt
#> 
#> # Unit test / coverage reports
#> htmlcov/
#> .tox/
#> .nox/
#> .coverage
#> .coverage.*
#> .cache
#> nosetests.xml
#> coverage.xml
#> *.cover
#> *.py,cover
#> .hypothesis/
#> .pytest_cache/
#> pytestdebug.log
#> 
#> # Translations
#> *.mo
#> *.pot
#> 
#> # Django stuff:
#> *.log
#> local_settings.py
#> db.sqlite3
#> db.sqlite3-journal
#> 
#> # Flask stuff:
#> instance/
#> .webassets-cache
#> 
#> # Scrapy stuff:
#> .scrapy
#> 
#> # Sphinx documentation
#> docs/_build/
#> doc/_build/
#> 
#> # PyBuilder
#> target/
#> 
#> # Jupyter Notebook
#> 
#> # IPython
#> 
#> # pyenv
#> .python-version
#> 
#> # pipenv
#> #   According to pypa/pipenv#598, it is recommended to include Pipfile.lock in version control.
#> #   However, in case of collaboration, if having platform-specific dependencies or dependencies
#> #   having no cross-platform support, pipenv may install dependencies that don't work, or not
#> #   install all needed dependencies.
#> #Pipfile.lock
#> 
#> # poetry
#> #poetry.lock
#> 
#> # PEP 582; used by e.g. github.com/David-OConnor/pyflow
#> __pypackages__/
#> 
#> # Celery stuff
#> celerybeat-schedule
#> celerybeat.pid
#> 
#> # SageMath parsed files
#> *.sage.py
#> 
#> # Environments
#> # .env
#> .env/
#> .venv/
#> env/
#> venv/
#> ENV/
#> env.bak/
#> venv.bak/
#> pythonenv*
#> 
#> # Spyder project settings
#> .spyderproject
#> .spyproject
#> 
#> # Rope project settings
#> .ropeproject
#> 
#> # mkdocs documentation
#> /site
#> 
#> # mypy
#> .mypy_cache/
#> .dmypy.json
#> dmypy.json
#> 
#> # Pyre type checker
#> .pyre/
#> 
#> # pytype static type analyzer
#> .pytype/
#> 
#> # operating system-related files
#> *.DS_Store #file properties cache/storage on macOS
#> Thumbs.db #thumbnail cache on Windows
#> 
#> # profiling data
#> .prof
#> 
#> 
#> ### R ###
#> # History files
#> .Rhistory
#> .Rapp.history
#> 
#> # Session Data files
#> .RData
#> 
#> # User-specific files
#> .Ruserdata
#> 
#> # Example code in package build process
#> *-Ex.R
#> 
#> # Output files from R CMD build
#> /*.tar.gz
#> 
#> # Output files from R CMD check
#> /*.Rcheck/
#> 
#> # RStudio files
#> .Rproj.user/
#> 
#> # produced vignettes
#> vignettes/*.html
#> vignettes/*.pdf
#> 
#> # OAuth2 token, see https://github.com/hadley/httr/releases/tag/v0.3
#> .httr-oauth
#> 
#> # knitr and R markdown default cache directories
#> *_cache/
#> /cache/
#> 
#> # Temporary files created by R markdown
#> *.utf8.md
#> *.knit.md
#> 
#> # R Environment Variables
#> .Renviron
#> 
#> ### R.Bookdown Stack ###
#> # R package: bookdown caching files
#> /*_files/
#> 
#> ### vscode ###
#> .vscode/*
#> !.vscode/settings.json
#> !.vscode/tasks.json
#> !.vscode/launch.json
#> !.vscode/extensions.json
#> *.code-workspace
#> 
#> # End of https://www.toptal.com/developers/gitignore/api/r,python,jupyternotebooks,vscode
#> 
#> ### targets ###
#> _targets
#> */out/*
#> !*.empty
#> 
#> ### vim ###
#> .*.swp
#> .swp
#> excluded_file.R
#> excluded_dir
#> *excluded_pattern*