Set up a USGS project directory

Wraps the other use_*_usgs functions to initialize multiple files in one directory.

use_project_usgs(
  home = ".",
  gitignore_additions = NULL,
  readme_rmd = FALSE,
  disclaimer_approved = FALSE,
  use_mr_template = TRUE
)

Arguments

home

chr, root directory of project. Defaults to current working directory

gitignore_additions

chr vector, other files/directories to be added to .gitignore

readme_rmd

lgl, should a README.Rmd file be created? If not (default), then a README.md is created.

disclaimer_approved

lgl, should this project contain an approved disclaimer statement? If not (default), then a provisional disclaimer statement is created.

use_mr_template

lgl, should this project contain the file necessary to have a default GitLab Merge Request template? It is included by default.

See also

Examples

tmp <- tempdir()
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

use_project_usgs(home = tmp)
#> ℹ Using README.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using DISCLAIMER_PROVISIONAL.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CHANGELOG.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CONTRIBUTING.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CODE_OF_CONDUCT.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using .gitignore from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
list.files(tmp)
#> [1] "CHANGELOG.md"              "CODE_OF_CONDUCT.md"       
#> [3] "CONTRIBUTING.md"           "DISCLAIMER_PROVISIONAL.md"
#> [5] "README.md"                

# start over
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

# creates README.Rmd and DISCLAIMER_APPROVED instead
use_project_usgs(home = tmp, readme_rmd = TRUE, disclaimer_approved = TRUE)
#> ℹ Using README.Rmd from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ✔ Setting active project to 'REDACTED FILE PATH'
#> ✔ Writing '.git/hooks/pre-commit'
#> ℹ Using DISCLAIMER_APPROVED.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CHANGELOG.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CONTRIBUTING.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using CODE_OF_CONDUCT.md from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
#> ℹ Using .gitignore from
#>   <https://code.usgs.gov/water/IWAAs-trends/templates/standard-template-repository/>
list.files(tmp)
#> [1] "CHANGELOG.md"           "CODE_OF_CONDUCT.md"     "CONTRIBUTING.md"       
#> [4] "DISCLAIMER_APPROVED.md" "README.Rmd"