Initialize a targets project

Run this function in a new targets project directory. It will create R files and directories using the targets project structure we often use in the IIDD Data Science Branch (see https://code.usgs.gov/wma/dsp/trainings/ds-pipelines-targets-2-course for an example).

tar_init(
  phase_names,
  phase_nums = seq_along(phase_names),
  home = ".",
  separate_phase_scripts = TRUE,
  phase_subdirs = c("src", "out"),
  overwrite = FALSE
)

Arguments

phase_names

chr vector, names of target phases like "fetch", "process", etc.

phase_nums

int vector, numbers to prepend to phase_names like "1_fetch", "2_process", etc. Defaults to 1:length(phase_names)

home

chr, root directory of targets project. Defaults to current working directory "."

separate_phase_scripts

lgl, should a different R script be created for each phase like "1_fetch.R", "2_process.R", etc? If FALSE, then targets lists will be initialized in _targets.R file. Defaults to TRUE

phase_subdirs

chr vector, subdirectories within each phase. Defaults to "src", "out"

overwrite

lgl, should the initialization overwrite files and folders that already exist? Defaults to FALSE

Value

NULL invisibly

Examples

# temporary directories in which targets project is initialized (you can skip this part)
tmp <- tempdir()
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

# creates 1_fetch, 2_process, 3_summarize R scripts and directories
tar_init(home = tmp, phase_names = c("fetch", "process", "summarize"))

list.files(tmp, full.names = FALSE, recursive = TRUE, all.files = TRUE, pattern = "\\.(R|empty)$")
#>  [1] "1_fetch.R"              "1_fetch/.empty"         "1_fetch/out/.empty"    
#>  [4] "1_fetch/src/.empty"     "2_process.R"            "2_process/.empty"      
#>  [7] "2_process/out/.empty"   "2_process/src/.empty"   "3_summarize.R"         
#> [10] "3_summarize/.empty"     "3_summarize/out/.empty" "3_summarize/src/.empty"
#> [13] "_targets.R"            

# clean out tmp folder
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

# different structure starting with 0_config and including "in/" dir in each phase
tar_init(home = tmp,
         phase_names = c("config", "pull", "munge", "visualize"),
         phase_nums = 0:3,
         phase_subdirs = c("in", "src", "out"))

list.files(tmp, full.names = FALSE, recursive = TRUE, all.files = TRUE, pattern = "\\.(R|empty)$")
#>  [1] "0_config.R"             "0_config/.empty"        "0_config/in/.empty"    
#>  [4] "0_config/out/.empty"    "0_config/src/.empty"    "1_pull.R"              
#>  [7] "1_pull/.empty"          "1_pull/in/.empty"       "1_pull/out/.empty"     
#> [10] "1_pull/src/.empty"      "2_munge.R"              "2_munge/.empty"        
#> [13] "2_munge/in/.empty"      "2_munge/out/.empty"     "2_munge/src/.empty"    
#> [16] "3_visualize.R"          "3_visualize/.empty"     "3_visualize/in/.empty" 
#> [19] "3_visualize/out/.empty" "3_visualize/src/.empty" "_targets.R"