Set up a USGS project directory

Wraps the other use_*_usgs functions to initialize multiple files in one directory.

Usage

use_project_usgs(
  home = ".",
  gitignore_additions = NULL,
  readme_rmd = FALSE,
  disclaimer_approved = FALSE,
  open = rlang::is_interactive()
)

Arguments

home

chr, root directory of project. Defaults to current working directory

gitignore_additions

chr vector, other files/directories to be added to .gitignore

readme_rmd

lgl, should a README.Rmd file be created? If not (default), then a README.md is created. Using an Rmd file will allow you to include R code and output in your README.md.

disclaimer_approved

lgl, should this project contain an approved disclaimer statement? If not (default), then a provisional disclaimer statement is created.

open

lgl; whether to open the files for interactive editing

See also

Examples

if (FALSE) { # \dontrun{
tmp <- tempdir()
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

use_project_usgs(home = tmp)
list.files(tmp)

# start over
unlink(tmp, recursive = TRUE, force = TRUE)
dir.create(tmp)

# creates README.Rmd and DISCLAIMER_APPROVED instead
use_project_usgs(home = tmp, readme_rmd = TRUE, disclaimer_approved = TRUE)
list.files(tmp)
} # }